Rename relative_distance_sensor
