+++ s2e-ff/CMakeLists.txt
@@ -45,8 +45,8 @@ add_subdirectory(${S2E_CORE_DIR}/src/simulation s2e_core/simulation)
 
 set(SOURCE_FILES
   src/s2e_ff.cpp
-  src/components/aocs/RelativeDistanceSensor.cpp
-  src/components/aocs/InitializeRelativeDistanceSensor.cpp
+  src/components/aocs/relative_distance_sensor.cpp
+  src/components/aocs/initialize_relative_distance_sensor.cpp
   src/components/aocs/RelativePositionSensor.cpp
   src/components/aocs/InitializeRelativePositionSensor.cpp
   src/components/aocs/RelativeVelocitySensor.cpp
