+++ s2e-ff/src/components/aocs/initialize_relative_distance_sensor.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "RelativeDistanceSensor.hpp"
+#include "relative_distance_sensor.hpp"
 
 RelativeDistanceSensor InitializeRelativeDistanceSensor(ClockGenerator* clock_gen, const std::string file_name, const double compo_step_time_s,
                                                         const RelativeInformation& rel_info, const int reference_sat_id_input = -1);
